Fix key response time regression
Should handle key events (and gpio polling) after
the vsync wait instead of before.  If done before,
we lose up to 20ms of response time if after we
wake up from the sleep a key was pressed. Better to
dequeue after we wake up, right before we are going
to emulate the next frame.