Add fast fail / fast stat
After switching to fatfs, real stats on files were delaying
boot time.  This feature has the kernel read a bootstat.txt
file that tells it to either fast fail or fast stat certain
files to save boot time.  If bootstat.txt is not found, the
kernel falls back to a default list.