Fix audio sync issue at first boot
Resid was not synced to the live edge as closely as it should
have been immediately after boot.  I didn't notice until I loaded
the delay test instead of attaching a cartridge.  When the program is
loaded via prg, the delay is obvious.  Strange thing is if the user
simply opened the menu and stayed there for a few seconds, audio would
eventually catch up to the live edge.  But since audio should be properly
synced right off the bat, had to do something.  This is a hack to
get it synced by forcing resid out, putting fast sid in, then immediately
going back to resid.  Result is a little popping noise but ends up
syncing properly.