Fix bugs with OSD
Anything that changes ui_active status needs to be done after
control key is lifted so emulator gets the up event before we
transition.

Also, removed 'need_swap' variable. Always swapping and drawing whatever
is in the frame buffer so we don't need the emulator to change something
for us to draw.