Temp fix for 3.4, set all Pi3 to 60% passband
The arm clock check in viceemulatorcore.cpp does not work. It must
not be returning the real clock rate like.  The one in menu.c does
work (probably because it's later in the boot?) so there is a mismatch
between 60 and 90 passband freq.  This caused garbage noise.  Temporarily
setting all Pi3 to 60% until I can reliably get the clock or some other
way to know to go up to 90.