Restored lost negative padding
Somewhere along the lines I lost the ability to add pixels to the border
to expand the frame buffer larger than the visible area.  Restored this.
Min for hborder and vborder should have always been 0.  The extra padding
needs to be added to the available area, not subtracted from it.