Added dedicated reset menu
When using with a C64 case, the number of extra C=/CTRL + F-Key combos are
limited. So to have a more versatile approach one now can be configured to
select either hard or soft reset from an OSD menu, instead of having two
different combos, one for hard and one for soft reset.

Signed-off-by: Sven Oliver Moll <svolli@svolli.de>