Keyrah v2 keymap file for real c16 keyboard
A real c16 keyboard does not map well with
the provided default rpi_pos.vkm. Since it
is assumed users are building a C64 replica
from this project, rpi_pos.vkm will remain
the default.  But if users are replicating
a C16 inside a real C16 case using a keyrah,
they should switch to the c16 version.